requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
PyYAML==6.0.2
